move download_hf to submodule
