+++ src/hssm/distribution_utils/utils.py
@@ -1,0 +1,29 @@
+"""Utility functions for likelihoods."""
+
+from huggingface_hub import hf_hub_download
+
+REPO_ID = "franklab/HSSM"
+
+
+def download_hf(path: str):
+    """
+    Download a file from a HuggingFace repository.
+
+    Parameters
+    ----------
+    path : str
+        The path of the file to download in the repository.
+
+    Returns
+    -------
+    str
+        The local path where the file is downloaded.
+
+    Notes
+    -----
+    The repository is specified by the REPO_ID constant,
+    which should be a valid HuggingFace.co repository ID.
+    The file is downloaded using the HuggingFace Hub's
+     hf_hub_download function.
+    """
+    return hf_hub_download(repo_id=REPO_ID, filename=path)
